Extract create_flattened_cube_graph to graph_utils
Moves the function from field.py to a dedicated module so both
Field and SurfaceConfig can import it without a cross-dependency.

Adds a unit test verifying that every edge connects face-adjacent
cells in the flattened index space.
